+++ test/melon.service.ts
@@ -48,35 +48,6 @@ const melonSecurity = (MELON) => {
         },
       },
     },
-
-    rolesRights: {
-      super_admin: {
-        async defineCRUDAbility(can, cannot, ctx) {
-          can('crud', MELON);
-        },
-      },
-      admin: {},
-      trusted_user: {
-        maxBatchSize: 5,
-      },
-      user: {
-        async defineCRUDAbility(can, cannot, ctx) {
-          const user: CrudUser = ctx.user;
-          const userId = ctx.userId;
-          can('cud', MELON, { owner: userId });
-          cannot('cu', MELON, ['size']);
-          can('read', MELON);
-        },
-      },
-      guest: {
-        async defineCRUDAbility(can, cannot, ctx) {
-          can('read', MELON);
-        },
-      },
-    },
-
-    maxItemsPerUser: 10,
-    additionalItemsInDbPerTrustPoints: 1,
   } as CrudSecurity;
 };
 @Injectable()
