+++ test/src/services/melon/cmds.ts
@@ -1,0 +1,3 @@
+//Auto generated file
+
+export const serviceCmds = {};
