+++ test/src/services/melon/melon.service.spec.ts
@@ -1,27 +1,27 @@
 import { MongoDriver } from '@mikro-orm/mongodb';
 import { Test, TestingModule } from '@nestjs/testing';
-import { PictureService } from './picture.service';
+import { MelonService } from './melon.service';
 import { EICRUDModule } from '@eicrud/core';
 import { MikroOrmModule } from '@mikro-orm/nestjs';
-import Picture from './picture.entity';
+import Melon from './melon.entity';
 import { MyConfigService } from '../../eicrud.config.service';
 import { CRUD_CONFIG_KEY } from '@eicrud/core/config';
 
 describe('AppController', () => {
-  let myService: PictureService;
+  let myService: MelonService;
 
   beforeEach(async () => {
     const app: TestingModule = await Test.createTestingModule({
       imports: [
         MikroOrmModule.forRoot({
-          entities: [Picture],
+          entities: [Melon],
           driver: MongoDriver,
-          dbName: 'test-picture',
+          dbName: 'test-melon',
         }),
         EICRUDModule.forRoot(),
       ],
       providers: [
-        PictureService,
+        MelonService,
         {
           provide: CRUD_CONFIG_KEY,
           useClass: MyConfigService,
@@ -29,7 +29,7 @@ describe('AppController', () => {
       ],
     }).compile();
 
-    myService = app.get<PictureService>(PictureService);
+    myService = app.get<MelonService>(MelonService);
   });
 
   describe('root', () => {
