+++ test/src/services/melon/melon.service.ts
@@ -1,0 +1,17 @@
+import { ModuleRef } from '@nestjs/core';
+import Melon from './melon.entity';
+import { Injectable } from '@nestjs/common';
+import { getSecurity } from './melon.security';
+import { CrudService } from '@eicrud/core/crud';
+import { serviceCmds } from './cmds';
+import { CrudContext } from '@eicrud/core/crud';
+
+@Injectable()
+export class MelonService extends CrudService<Melon> {
+  constructor(protected moduleRef: ModuleRef) {
+    const serviceName = CrudService.getName(Melon);
+    super(moduleRef, Melon, getSecurity(serviceName));
+  }
+
+  // GENERATED START - do not remove
+}
