+++ test/src/services/userprofile/cmds.ts
@@ -1,3 +1,4 @@
+import { searchSecurity } from './cmds/search/search.security';
 import { test_cmd_rate_limitedSecurity } from './cmds/test_cmd_rate_limited/test_cmd_rate_limited.security';
 import { canCannotCmdSecurity } from './cmds/can_cannot_cmd/can_cannot_cmd.security';
 import { test_cmdSecurity } from './cmds/test_cmd/test_cmd.security';
@@ -5,6 +6,7 @@ import { test_cmdSecurity } from './cmds/test_cmd/test_cmd.security';
 //Auto generated file
 
 export const serviceCmds = {
+  search: searchSecurity,
   test_cmd_rate_limited: test_cmd_rate_limitedSecurity,
   can_cannot_cmd: canCannotCmdSecurity,
   test_cmd: test_cmdSecurity,
